requirements.txt
python-dotenv
urllib3>=1.26.0,<2.0.0 # Added version for better dependency management
pytest